modify : division code 타입 재설정

+++ src/main/java/com/hollroom/monthly/dao/MonthlyProductDAO.java
@@ -9,5 +9,5 @@ import java.util.List;
 public interface MonthlyProductDAO {
     void insertProduct(MonthlyProductEntity entity);
     List<MonthlyProductEntity> readProductAll();
-    List<MonthlyProductEntity> readDivisionProduct(int divisionCode);
+    List<MonthlyProductEntity> readDivisionProduct(Long divisionCode);
 }
