+++ src/main/java/com/hollroom/monthly/repository/MonthlyProductRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface MonthlyProductRepository extends JpaRepository<MonthlyProductEntity,Long> {
-    List<MonthlyProductEntity> findByDivisionCode(int divisionCode);
+    List<MonthlyProductEntity> findByDivisionCode(Long divisionCode);
 }
